+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Refactored `ErrorStack` class into `Errors` class.
 - Deprecated several error methods.
+- Improved validation error messaging.
 
 ### Fixed
 - Giving an invalid ID to `find()` no longer triggers a PHP error.